* bsd-kvm.c (bsd_kvm_pcb_cmd): Cast return value from parse_and_eval_address to u_long before converting it to a pointer type. * i386fbsd-nat.c (_initialize_i386fbsd_nat): Change type of ps_strings from int to u_long. * i386obsd-nat.c (_initialize_i386obsd_nat): Cast _ps.val to u_long instead of CORE_ADDR.

+++ gdb/bsd-kvm.c
@@ -259,7 +259,7 @@ bsd_kvm_pcb_cmd (char *arg, int fromtty)
   if (core_kd == NULL)
     error ("No kernel memory image.");
 
-  bsd_kvm_paddr = (struct pcb *) parse_and_eval_address (arg);
+  bsd_kvm_paddr = (struct pcb *)(u_long) parse_and_eval_address (arg);
 
   target_fetch_registers (-1);
